[tests] Extract out the mojo rule to its own class to assist in migration to junit 5
